requirements.txt
# --- Core ---
Flask==3.1.1
flask-cors==6.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
python-dotenv==1.0.1

# --- DB ---
SQLAlchemy==2.0.41
greenlet==3.2.4

# --- OpenAI (stable combo) ---
openai==1.42.0
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4

# --- HTTP stack compatible (for SSE keepalive, requests helpers) ---
httpx==0.27.2
httpcore==1.0.5
h11==0.14.0
idna==3.10
certifi==2025.8.3
urllib3==2.2.3
sniffio==1.3.1

# --- Misc utils (optional but used) ---
requests==2.32.3
tqdm==4.67.1

# --- Deployment (optional) ---
gunicorn==21.2.0